backend/requirements.txt
Django~=3.2.7
psycopg2~=2.9.1
Pillow~=8.3.2
requests~=2.26.0
djangorestframework~=3.12.4
Markdown~=3.3.4
django-filter~=21.1
djangorestframework-simplejwt~=4.8.0
django-cors-headers~=3.10.0
PyJWT==2.1.0